chore: reduce circular dependencies
