+++ packages/grafana-data/src/field/fieldColor.ts
@@ -13,13 +13,11 @@ import { t } from '@grafana/i18n';
 
 import { getContrastRatio } from '../themes/colorManipulator';
 import { type GrafanaTheme2 } from '../themes/types';
-import { reduceField } from '../transformations/fieldReducer';
 import { type Field } from '../types/dataFrame';
 import { FALLBACK_COLOR, FieldColorModeId } from '../types/fieldColor';
 import { type Threshold } from '../types/thresholds';
 import { Registry, type RegistryItem } from '../utils/Registry';
 
-import { getScaleCalculator, type ColorScaleValue } from './scale';
 import { fallBackThreshold } from './thresholds';
 
 /**
@@ -360,30 +358,6 @@ export function getFieldColorModeForField(field: Field): FieldColorMode {
 /** @beta */
 export function getFieldColorMode(mode?: FieldColorModeId | string): FieldColorMode {
   return fieldColorModeRegistry.getIfExists(mode) ?? fieldColorModeRegistry.get(FieldColorModeId.Thresholds);
-}
-
-/**
- * @alpha
- * Function that will return a series color for any given color mode. If the color mode is a by value color
- * mode it will use the field.config.color.seriesBy property to figure out which value to use
- */
-export function getFieldSeriesColor(field: Field, theme: GrafanaTheme2): ColorScaleValue {
-  const mode = getFieldColorModeForField(field);
-
-  if (!mode.isByValue) {
-    return {
-      color: mode.getCalculator(field, theme)(0, 0),
-      threshold: fallBackThreshold,
-      percent: 1,
-    };
-  }
-
-  const scale = getScaleCalculator(field, theme);
-  const stat = field.config.color?.seriesBy ?? 'last';
-  const calcs = reduceField({ field, reducers: [stat] });
-  const value = calcs[stat] ?? 0;
-
-  return scale(value);
 }
 
 export function getColorByStringHash(colors: string[], string: string) {
