+++ packages/grafana-data/src/field/fieldState.ts
@@ -1,4 +1,3 @@
-import { getFieldMatcher } from '../transformations/matchers';
 import { type Labels } from '../types/data';
 import {
   type DataFrame,
@@ -7,7 +6,6 @@ import {
   TIME_SERIES_TIME_FIELD_NAME,
   TIME_SERIES_VALUE_FIELD_NAME,
 } from '../types/dataFrame';
-import { type FieldConfigSource } from '../types/fieldOverrides';
 import { findUniqueLabels, formatLabels } from '../utils/labels';
 
 /**
@@ -59,54 +57,6 @@ export function cacheFieldDisplayNames(frames: DataFrame[]) {
       if (field.type === FieldType.nestedFrames) {
         field.values.forEach(cacheFieldDisplayNames);
       }
-    });
-  });
-}
-
-/**
- *
- * moves each field's config.custom.hideFrom to field.state.hideFrom
- * and mutates original field.config.custom.hideFrom to one with explicit overrides only, (without the ad-hoc stateful __system override from legend toggle)
- */
-export function decoupleHideFromState(frames: DataFrame[], fieldConfig: FieldConfigSource) {
-  frames.forEach((frame) => {
-    frame.fields.forEach((field) => {
-      const hideFrom = {
-        legend: false,
-        tooltip: false,
-        viz: false,
-        ...fieldConfig.defaults.custom?.hideFrom,
-      };
-
-      // with ad hoc __system override applied
-      const hideFromState = field.config.custom?.hideFrom;
-
-      fieldConfig.overrides.forEach((o) => {
-        if ('__systemRef' in o) {
-          return;
-        }
-
-        const m = getFieldMatcher(o.matcher);
-
-        if (m(field, frame, frames)) {
-          for (const p of o.properties) {
-            if (p.id === 'custom.hideFrom') {
-              Object.assign(hideFrom, p.value);
-            }
-          }
-        }
-      });
-
-      field.state = {
-        ...field.state,
-        hideFrom: {
-          ...hideFromState,
-        },
-      };
-
-      // original with perm overrides
-      field.config.custom = field.config.custom ?? {};
-      field.config.custom.hideFrom = hideFrom;
     });
   });
 }
