+++ packages/grafana-data/src/internal/index.ts
@@ -71,7 +71,8 @@ export {
 } from '../transformations/transformers/calculateField';
 export { type SeriesToRowsTransformerOptions } from '../transformations/transformers/seriesToRows';
 export { histogramFieldInfo, type HistogramTransformerInputs } from '../transformations/transformers/histogram';
-export { joinByFieldTransformer, type JoinByFieldOptions, JoinMode } from '../transformations/transformers/joinByField';
+export { joinByFieldTransformer, type JoinByFieldOptions } from '../transformations/transformers/joinByField';
+export { JoinMode } from '../transformations/transformers/joinShared';
 export { LabelsToFieldsMode, type LabelsToFieldsOptions } from '../transformations/transformers/labelsToFields';
 export { type LimitTransformerOptions } from '../transformations/transformers/limit';
 export { type MergeTransformerOptions } from '../transformations/transformers/merge';
@@ -116,7 +117,7 @@ export { type PluginAddedLinksConfigureFunc, type PluginExtensionEventHelpers } 
 
 export { getStreamingFrameOptions } from '../dataframe/StreamingDataFrame';
 export { fieldIndexComparer } from '../field/fieldComparers';
-export { decoupleHideFromState } from '../field/fieldState';
+export { decoupleHideFromState } from '../field/decoupleHideFromState';
 export { findNumericFieldMinMax } from '../field/fieldOverrides';
 export { type PanelOptionsSupplier } from '../panel/PanelPlugin';
 export { sanitize, sanitizeUrl } from '../text/sanitize';
