+++ packages/grafana-data/src/transformations/transformers/joinByField.ts
@@ -11,12 +11,7 @@ import { FieldMatcherID } from '../matchers/ids';
 
 import { DataTransformerID } from './ids';
 import { joinDataFrames } from './joinDataFrames';
-
-export enum JoinMode {
-  outer = 'outer', // best for time series, non duplicated join on values
-  inner = 'inner',
-  outerTabular = 'outerTabular', // best for tabular data where the join on value can be duplicated
-}
+import { JoinMode } from './joinShared';
 
 export interface JoinByFieldOptions {
   byField?: string; // empty will pick the field automatically
