+++ packages/grafana-data/src/transformations/transformers/joinShared.ts
@@ -1,0 +1,5 @@
+export enum JoinMode {
+  outer = 'outer', // best for time series, non duplicated join on values
+  inner = 'inner',
+  outerTabular = 'outerTabular', // best for tabular data where the join on value can be duplicated
+}
