+++ packages/grafana-data/src/types/displayValue.ts
@@ -1,4 +1,4 @@
-import { type FormattedValue } from '../valueFormats/valueFormats';
+import { type FormattedValue } from './valueFormats';
 
 export type DisplayProcessor = (value: unknown, decimals?: DecimalCount) => DisplayValue;
