+++ packages/grafana-data/src/utils/csv.ts
@@ -7,7 +7,7 @@ import { MutableDataFrame } from '../dataframe/MutableDataFrame';
 import { guessFieldTypeFromValue } from '../dataframe/guessFieldType';
 import { getFieldDisplayName } from '../field/fieldState';
 import { type DataFrame, type Field, type FieldConfig, FieldType } from '../types/dataFrame';
-import { formattedValueToString } from '../valueFormats/valueFormats';
+import { formattedValueToString } from '../valueFormats/baseFormatters';
 
 export enum CSVHeaderStyle {
   full,
