+++ packages/grafana-data/src/valueFormats/arithmeticFormatters.ts
@@ -1,6 +1,7 @@
 import { type DecimalCount } from '../types/displayValue';
+import { type FormattedValue } from '../types/valueFormats';
 
-import { toFixed, type FormattedValue } from './valueFormats';
+import { toFixed } from './baseFormatters';
 
 export function toPercent(size: number | null, decimals: DecimalCount): FormattedValue {
   if (size === null) {
